TestCase code: simplified the code and improved the functionality of test case automation with the below (new) features of the aeunittest library 1) loading the text cases csv file instead of hardcoding the test cases
also some code cleanup and fixing a statuscode for a Goal endpoint
