+++ testcases/lifelog test cases - Activity.csv
@@ -1,0 +1,35 @@
+Name,Uri,HttpVerb,RequestBody,WantStatusCode,WantResponseBody,Skip,ref #,Notes
+Activity Post Field validation-String field,/activity,POST,"{""Name"":123,""GoalID"":""Test123""}",400 bad request,?,,1.a.1,
+Activity Post Field validation-String field,/activity,POST,"{""Name"":""123"",""GoalID"":123}",400 bad request,?,,1.a.2,
+Activity Post Field validation-Passing blank in mandatory fields,/activity,POST,"{""Name"":"""".""GoalID"":""test4""}",400 bad request,?,,1.b,
+Activity Post new record test,/activity,POST,"{""Name"":""test1"",""GoalID"":""test""}",201 (created),"{""Name"":""test1"",""GoalID"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,2.1,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+Activity Post Get record,/activity/test1,GET,,200 ok,,,2.2,
+Activity Post new record with duplicate key ,/activity,POST,"{""Name"":""test1"",""GoalID"":""test""}",400 bad request,"{""Name"":""test1"",""GoalID"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,3,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+Activity Post with typo in field name in json/Sending Extra field,/activity,POST,"{""Name"":""test2"",""ID"":""test2""}",400 bad request,?,,4,
+Activity Post Invalid Json,/activity,POST,"{""Name"":""test3""GoalID"":""testGoalID5""}",404 not found,?,,5,
+Activity Post without body,/activity,POST,,400 bad request,?,,6,
+Activity Post with wrong key name,/activity,POST,"{""id"":""test4"",""GoalID"":""testGoalID6""}",400 bad request,,,7,
+,,,,,,,,
+Activity Get with key,/activity/test1,GET,,200 ok,"check if all the fields are correct in the response as per the database, including the key",,1.a,
+Activity Get time field in response should be of RFC 3339 format ,/activity/test1,GET,,200 ok ,"example “2016-10-02T15:04:05-07:00""",,1.b,
+Activity Get with nonexisting Key,/activity/xyz,GET,,404 not found,,,2,
+Activity Get with blank key,/activity/,GET,,200 ok ,,,3,
+,,,,,,,,
+Activity Put Field validation-String field,/activity/test1,PUT,"{""Name"":""test1"",""GoalID"":123}",406 not acceptable,?,,1.a,
+Activity Put Field validation-Passing blank in mandatory fields,/activity,PUT,"{""Name"":"""",""GoalID"":""test4""}",406 not acceptable,?,,1.b,
+Activity Put Modifying a record,/activity/test1,PUT,"{""Name"":""test1"",""GoalID"":""TestGoalID New""}",200 ok,,,2.1,
+Activity Put Successful saving of the record to database,/activity/test1,GET,,200 ok,"{""Name"":""test1"",""GoalID"":""TestGoalID New""}",,2.2,
+Activity Put new key should NOT be allowed,/activity/xyz,PUT,"{""Name"":""xyz"",""GoalID"":""TestGoalID New""}",400 bad request,,,3,
+Activity Put with typo in field name in json/Sending Extra field,/activity/test1,PUT,"{""Name"":""test1"",""Note"":""test2""}",400 bad request,?,,4,
+Activity Put Invalid Json,/activity/test1,PUT,"{""Name"":""test1""GoalID"":""testGoalID3""}",404 not found,?,,5,
+Activity Put key in URL not same as key in Body,/activity/test1,PUT,"{""Name"":""test10"",""GoalID"":""Test123""}",400 bad request,,,6,
+Activity Put without parameter in URI,/activity/,PUT,"{""Name"":""test1"",""GoalID"":""Test1New""}",403 forbidden,,,7,
+,,,,,,,,
+Activity Delete without key,/activity/,DELETE,,403 forbidden,,,1,
+Activity Delete remove record,/activity/test1,DELETE,,200 ok,,,2.1,
+Activity Delete Get same record,/activity/test1,GET,,404 not found,,,2.2,
+Activity Delete  non Existing record,/activity/xyz,DELETE,,404 not found,,,3,
+,,,,,,,,
+#Notes from testing:,,,,,,,,
+#1) order of test cases in this document is not same as order of test records in code,,,,,,,,
+#2) there is a case in code 'Activity Post duplicate record test' that is not documented here. looks like it is already covered in excel but with a different name.,,,,,,,,
