+++ testcases/lifelog test cases - ActivityLog.csv
@@ -1,0 +1,34 @@
+Name,Uri,HttpVerb,RequestBody,WantStatusCode,WantResponseBody,Skip,ref #,Notes
+ActivityLog Post Field validation-String field,/activitylog,POST,"{""Name"":123,""Notes"":""Test123""}",400 bad request,?,,1.a.1,
+ActivityLog Post Field validation-String field,/activitylog,POST,"{""Name"":""123"",""Notes"":123}",400 bad request,?,,1.a.2,
+ActivityLog Post Field validation-Passing blank in mandatory fields,/activitylog,POST,"{""Name"":"""".""Notes"":""test4""}",400 bad request,?,,1.b,
+ActivityLog Post Field Validation-Date fields validation,/activitylog,POST,"{""Name"":""test1"",""Notes"":""test5""}",400 bad request,,,1.c.1,test for start date field
+ActivityLog Post Field Validation-Date fields validation,/activitylog,POST,"{""Name"":""test1"",""Notes"":""test5""}",400 bad request,,,1.c.2,test for end date field
+ActivityLog Post new record test,/activitylog,POST,"{""Name"":""test1"",""Notes"":""test""}",201 (created),"{""Name"":""test1"",""Notes"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,2,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+ActivityLog Post Get same record,/activitylog/test1,GET,,200 ok,,,2.2,
+ActivityLog Post new record with duplicate key ,/activitylog,POST,"{""Name"":""test1"",""Notes"":""test""}",400 bad request,"{""Name"":""test1"",""Notes"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,3,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+ActivityLog Post with typo in field name in json/Sending Extra field,/activitylog,POST,"{""Name"":""test2"",""Note"":""test2""}",400 bad request,?,TRUE,4,
+ActivityLog Post Invalid Json,/activitylog,POST,"{""Name"":""test3""Notes"":""testnotes3""}",404 not found,?,,5,
+ActivityLog Post without body,/activitylog,POST,,400 bad request,?,,6,
+ActivityLog Post with wrong key name,/activitylog,POST,"{""id"":""test4"",""Notes"":""testNotes4""}",400 bad request,,TRUE,7,
+,,,,,,,,
+ActivityLog Get with key,/activitylog/test1,GET,,200 ok,"check if all the fields are correct in the response as per the database, including the key",,1.a,
+ActivityLog Get time field in response should be of RFC 3339 format ,/activitylog/test1,GET,,200 ok ,"example “2016-10-02T15:04:05-07:00""",,1.b,
+ActivityLog Get with nonexisting Key,/activitylog/xyz,GET,,404 not found,,,2,
+ActivityLog Get with blank key,/activitylog/,GET,,200 OK,,,3,
+,,,,,,,,
+ActivityLog Put Field validation-String field,/activitylog/test1,PUT,"{""Name"":""test1"",""Notes"":123}",400 bad request,?,,1.a,
+ActivityLog Put Field validation-Passing blank in mandatory fields,/activitylog,PUT,"{""Name"":"""",""Notes"":""test4""}",400 bad request,?,,1.b,
+ActivityLog Put Field Validation-Date fields validation,/activitylog,PUT,,400 bad request,,,1.c,
+ActivityLog Put Modifying a record,/activitylog/test1,PUT,"{""Name"":""test1"",""Notes"":""TestNotes New""}",200 ok,,,2.1,
+ActivityLog Put Successful saving of the record to database,/activitylog/test1,GET,,200 ok,"{""Name"":""test1"",""Notes"":""TestNotes New""}",,2.2,
+ActivityLog Put new key should NOT be allowed,/activitylog/xyz,PUT,"{""Name""xyz"":""Notes"",""TestNotes New""}",400 bad request,,,3,
+ActivityLog Put with typo in field name in json/Sending Extra field,/activitylog/test1,PUT,"{""Name"":""test1"",""Note"":""test2""}",400 bad request,?,,4,
+ActivityLog Put Invalid Json,/activitylog/test1,PUT,"{""Name"":""test1""Notes"":""testnotes3""}",404 not found,?,,5,
+Activity Put key in URL not same as key in Body,/activitylog/test1,PUT,"{""Name"":""test10"",""Notes"":""Test123""}",400 bad request,,,6,
+Activity Put without parameter in URI,/activitylog/test1,PUT,"{""Name"":""test1"",""Notes"":123}",403 forbidden,,,7,
+,,,,,,,,
+ActivityLog Delete without key,/activitylog/,DELETE,,403 forbidden,,,1,
+ActivityLog Delete remove record,/activitylog/test1,DELETE,,200 ok,,,2.1,
+ActivityLog Delete Get same record,/activitylog/test1,GET,,200 ok,,,2.2,
+ActivityLog Delete  non Existing record,/activitylog/xyz,DELETE,,404 not found,,,3,
