+++ testcases/lifelog test cases - Goal.csv
@@ -1,0 +1,35 @@
+Name,Uri,HttpVerb,RequestBody,WantStatusCode,WantResponseBody,Skip,ref #,Notes
+Goal Post Field validation-String field,/goal,POST,"{""Name"":123,""Notes"":""Test123""}",400 bad request,?,,1.a.1,
+Goal Post Field validation-String field,/goal,POST,"{""Name"":""123"",""Notes"":123}",400 bad request,?,,1.a.2,
+Goal Post Field validation-Passing blank in mandatory fields,/goal,POST,"{""Name"":"""".""Notes"":""test4""}",400 bad request,?,,1.b,
+Goal Post new record test,/goal,POST,"{""Name"":""test1"",""Notes"":""test1""}",201 (created),"{""Name"":""test1"",""Notes"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,2.1,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+Goal Post Get record,/goal/test1,GET,,200 ok,,,2.2,
+Goal Post new record with duplicate key ,/goal,POST,"{""Name"":""test1"",""Notes"":""test1""}",400 bad request,"{""Name"":""test1"",""Notes"":""test"",""CreatedOn"":"""",""ModifiedOn"":""""}",,3,"the value for the date fields is ignored in ""want - Response body"" as they are not deterministic."
+Goal Post with typo in field name in json/Sending Extra field,/goal,POST,"{""Name"":""test2"",""Note"":""test2""}",400 bad request,?,,4,
+Goal Post Invalid Json,/goal,POST,"{""Name"":""test3""Notes"":""testnotes3""}",400 bad request,?,,5,
+Goal Post without body,/goal,POST,,400 bad request,?,,6,
+Goal Post with wrong key name,/goal,POST,"{""id"":""test4"",""Notes"":""testNotes4""}",400 bad request,,,7,
+,,,,,,,,
+Goal Get with key,/goal/test1,GET,,200 ok,"check if all the fields are correct in the response as per the database, including the key",,1.a,
+Goal Get time field in response should be of RFC 3339 format ,/goal/test1,GET,,200 ok ,"example “2016-10-02T15:04:05-07:00""",,1.b,
+Goal Get with nonexisting Key,/goal/xyz,GET,,404 not found,,,2,
+Goal Get with blank key,/goal/,GET,,200 ok ,,,3,
+,,,,,,,,
+Goal Put Field validation-String field,/goal/test1,PUT,"{""Name"":""test1"",""Notes"":123}",400 bad request,?,,1.a,
+Goal Put Field validation-Passing blank in mandatory fields,/goal/test1,PUT,"{""Name"":"""",""Notes"":""test4""}",400 bad request,?,,1.b,
+Goal Put Editing a record,/goal/test1,PUT,"{""Name"":""test1"",'""Notes"":""TestNotes New""}",200 ok,,,2.1,
+Goal Put Successful saving of the record to database,/goal/test1,GET,,200 ok,"{""Name"":""test1"",'""Notes"":""TestNotes New""}",,2.2,
+Goal Put new key should NOT be allowed,/goal/xyz,PUT,"{""Name"":""xyz"",""Notes"":""Notes123""}",400 bad request,,,3,
+Goal Put with typo in field name in json/Sending Extra field,/goal/test1,PUT,"{""Name"":""test1"",""Note"":""test2""}",400 bad request,?,,4,
+Goal Put Invalid Json,/goal/test1,PUT,"{""Name"":""test3""Notes"":""testnotes3""}",400 bad request,?,,5,
+Goal Put key in URL not same as key in Body,/goal/test1,PUT,"{""Name"":""test10"",""Notes"":""Test123""}",400 bad request,,,6,
+Goal Put without parameter in URI,/goal/,PUT,"{""Name"":""test1"",""Notes"":""test1new""}",403 forbidden,,,7,
+,,,,,,,,
+Goal Delete without key,/goal/,DELETE,,403 forbidden,,,1,
+Goal Delete remove record,/goal/test1,DELETE,,200 ok,,,2.1,
+Goal Delete Get same record,/goal/test1,GET,,404 not found,,,2.2,
+Goal Delete non Existing record,/goal/xyz,DELETE,,404 not found,,,3,
+,,,,,,,,
+,,,,,,,,
+#Notes from testing:,,,,,,,,
+#1) order of test cases in this document is not same as order of test records in code,,,,,,,,
